Get values from QVariant before comparing
Also clean up some minor issues

+++ rpcs3/rpcs3qt/game_list.h
@@ -12,7 +12,7 @@ class game_list : public QTableWidget
 private:
 	void mousePressEvent(QMouseEvent *event) override
 	{
-		if (!indexAt(event->pos()).isValid() || itemAt(event->pos())->data(Qt::UserRole) < 0)
+		if (!indexAt(event->pos()).isValid() || itemAt(event->pos())->data(Qt::UserRole).toInt() < 0)
 		{
 			clearSelection();
 			setCurrentItem(nullptr); // Needed for currentItemChanged
